requirements.txt
# web framework
flask

# Azure SDK components
azure-cognitiveservices-vision-computervision

# simplified system variables
python-dotenv